Remove reference to net/url URL.RawFragment.
RawFragment was introduced only in go1.15:
https://golang.org/doc/go1.15#net/url

If RawFragment works like RawPath and RawQuery, then RawFragment being
set implies Fragment is set, so the `if` conditional should have the
same effect. In returning the composed URL, the worst that may happen is
we'll lose the exact original escaping of the fragment.