Don't terminate noiseLoop when noise.AcceptHandshake fails.
AcceptHandshake can fail for reasons as simple as a message being
incorrectly formatted. We only want to exit this main processing loop
when an unrecoverable error occurs.

Before this change, a client could cause a server to stop processing
messages by sending it a message consisting only of "\x01". noiseLoop
would exit with:
noise: message is too short